+++ Changelog.md
@@ -3,4 +3,5 @@
 - Merged #65 from @sukibaby, which fixes scripts not launching if spaces are in the source path
 - Merged #66 from @angeld7, which improves functionality of the select all box
 - Optimized regex pattern for faster device identification
-
+- Updated dependencies
+- Changed device state names (Not Detected, Generic ADB, and $devicename)